Added support for AsciiDoc format in the narrative texts.
You can now use AsciiDoc in the narrative texts. To activate this option, set the narrative.format system property to "asciidoc".

+++ thucydides-core/src/main/java/net/thucydides/core/reports/html/HtmlReporter.java
@@ -80,7 +80,7 @@ public abstract class HtmlReporter extends ThucydidesReporter {
 
     protected void copyTestResultsToOutputDirectory() throws IOException {
         File testResultsSource = getSourceDirectoryOrDefault();
-        if ((!getOutputDirectory().equals(testResultsSource))) {
+        if ((!getOutputDirectory().equals(testResultsSource)) && testResultsSource.exists()) {
             FileUtils.copyDirectory(testResultsSource, getOutputDirectory(), withXMLorHTMLorCSVFiles());
         }
     }
